requirements.txt
alembic==1.13.1
attrs==23.2.0
blinker==1.7.0
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-marshmallow==1.2.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.20.2
packaging==23.2
platformdirs==4.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
requests-cache==1.1.1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
url-normalize==1.4.3
urllib3==2.2.0
Werkzeug==3.0.1
WTForms==3.1.2
